Split out thread/semaphore into their own header and add a guard/lock class.
